serving/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
torch
torchvision
scikit-learn
python-multipart
opencv-python-headless